+++ litho-widget-kotlin/src/main/kotlin/com/facebook/litho/widget/collection/CollectionRecyclerSpec.kt
@@ -157,7 +157,6 @@ object CollectionRecyclerSpec {
                 isCircular(isCircular)
                 hasDynamicItemHeight(hasDynamicItemHeight())
                 componentsConfiguration(componentsConfiguration)
-                canInterruptAndMoveLayoutsBetweenThreads(moveLayoutsBetweenThreads())
                 isReconciliationEnabled(isReconciliationEnabled)
                 isLayoutDiffingEnabled(isLayoutDiffingEnabled)
                 componentWarmer(componentWarmer)