+++ litho-widget/src/main/java/com/facebook/litho/widget/ComponentTreeHolder.java
@@ -52,7 +52,6 @@ import javax.annotation.concurrent.ThreadSafe;
 public class ComponentTreeHolder {
   private static final int UNINITIALIZED = -1;
   private static final AtomicInteger sIdGenerator = new AtomicInteger(1);
-  private final boolean mCanInterruptAndMoveLayoutsBetweenThreads;
   private final boolean mIsReconciliationEnabled;
   private final boolean mIsLayoutDiffingEnabled;
   public static final String PREVENT_RELEASE_TAG = "prevent_release";
@@ -130,7 +129,6 @@ public class ComponentTreeHolder {
     private @Nullable RunnableHandler preallocateMountContentHandler;
     private boolean shouldPreallocatePerMountSpec;
     private boolean incrementalMount = true;
-    private boolean canInterruptAndMoveLayoutsBetweenThreads;
     private boolean isReconciliationEnabled = ComponentsConfiguration.isReconciliationEnabled;
     private boolean isLayoutDiffingEnabled = ComponentsConfiguration.isLayoutDiffingEnabled;
     private int recyclingMode;
@@ -187,12 +185,6 @@ public class ComponentTreeHolder {
       return this;
     }
 
-    @Deprecated
-    public Builder canInterruptAndMoveLayoutsBetweenThreads(boolean isEnabled) {
-      this.canInterruptAndMoveLayoutsBetweenThreads = isEnabled;
-      return this;
-    }
-
     public Builder isReconciliationEnabled(boolean isEnabled) {
       isReconciliationEnabled = isEnabled;
       return this;
@@ -239,7 +231,6 @@ public class ComponentTreeHolder {
     mPreallocateMountContentHandler = builder.preallocateMountContentHandler;
     mShouldPreallocatePerMountSpec = builder.shouldPreallocatePerMountSpec;
     mComponentTreeMeasureListenerFactory = builder.componentTreeMeasureListenerFactory;
-    mCanInterruptAndMoveLayoutsBetweenThreads = builder.canInterruptAndMoveLayoutsBetweenThreads;
     mId = sIdGenerator.getAndIncrement();
     mIncrementalMount = builder.incrementalMount;
     mVisibilityProcessingEnabled = builder.visibilityProcessingEnabled;
@@ -491,7 +482,6 @@ public class ComponentTreeHolder {
                   : mComponentTreeMeasureListenerFactory.create(this))
           .incrementalMount(mIncrementalMount)
           .visibilityProcessing(mVisibilityProcessingEnabled)
-          .canInterruptAndMoveLayoutsBetweenThreads(mCanInterruptAndMoveLayoutsBetweenThreads)
           .logger(mRenderInfo.getComponentsLogger(), mRenderInfo.getLogTag())
           .componentsConfiguration(mComponentsConfiguration)
           .build();