+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 _release-date_
 
-* TBD
+* [BREAKING] Removed `ComponentsConfiguration.canInterruptAndMoveLayoutsBetweenThreads`, `ComponentTree.canInterruptAndMoveLayoutsBetweenThreads`, `RecyclerBinder.canInterruptAndMoveLayoutsBetweenThreads`, `ComponentTreeHolder.canInterruptAndMoveLayoutsBetweenThreads`. ComponentTrees can no longer exempt themselves from interrupt and move layouts between threads operation.
 
 For more details, see the [full diff](https://github.com/facebook/litho/compare/v0.47.0...master).